Add untested and hacky support for DVB-T2.
This patch adds *untested* support for DVB-T2. The patch is made
by reading Linux media API docs about T2 and no testing was done
with any hardware. The way to choose T2 is a hack and should be done
properly if this patch is working at all.

So, in order to instruct kernel to use T2 you have to give dvblast
one of the two options --guard -2 or --transmission -2 the other
options should be set like you are setting DVB-T.

PLP_ID is hardcoded to 0. There is not enough documentation for
this property so no aditional command line options were added.